feat(simd): stabilize kernels and dispatch table for amd64 and arm64
- Normalized all assembly stubs to use explicit signatures in all_kernels_gen.go.
- Fixed prefetchNTA to use uintptr across all platforms.
- Updated ImplementationDispatch to include missing transcendental and distance metrics.
- Added missing AVX2/NEON fallbacks for math functions.
- Resolved symbol redeclaration and type mismatch errors in go vet.

+++ internal/simd/coverage_gap_test.go
@@ -106,7 +106,7 @@ func TestCoverage_F16BatchStubs(t *testing.T) {
 
 func TestCoverage_Misc(t *testing.T) {
 	dummy := make([]byte, 16)
-	prefetchNTA(unsafe.Pointer(&dummy[0]))
+	prefetchNTA(uintptr(unsafe.Pointer(&dummy[0])))
 	
 	andBytesAVX2(dummy, dummy)
 	orBytesAVX2(dummy, dummy)
